requirements.txt
Django==2.2.14
djangorestframework-simplejwt==5.2.0
djangorestframework==3.13.1
python-decouple==3.1
django_jsonfield==1.4.1
awesome-slugify==1.6.5